+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - The lifetime of auto generated TLS certificates is now configurable with the role and roleGroup
   config property `requestedSecretLifetime`. This helps reducing frequent Pod restarts ([#619]).
+- Run a `containerdebug` process in the background of each HDFS container to collect debugging information ([#629]).
 
 ### Fixed
 
@@ -17,6 +18,7 @@ All notable changes to this project will be documented in this file.
 
 [#616]: https://github.com/stackabletech/hdfs-operator/pull/616
 [#619]: https://github.com/stackabletech/hdfs-operator/pull/619
+[#629]: https://github.com/stackabletech/hdfs-operator/pull/629
 
 ## [24.11.0] - 2024-11-18